clean up test module structure and pytest hooks
