+++ unit_tests/resources/source_pokeapi_w_components_py/components.py
@@ -1,7 +1,7 @@
 """A sample implementation of custom components that does nothing but will cause syncs to fail if missing."""
 
 from collections.abc import Iterable, MutableMapping
-from typing import Any, Mapping
+from typing import Any
 
 import requests
