+++ unit_tests/source_declarative_manifest/test_source_declarative_w_custom_components.py
@@ -32,7 +32,7 @@ from airbyte_cdk.sources.declarative.parsers.custom_code_compiler import (
     custom_code_execution_permitted,
     register_components_module_from_string,
 )
-from airbyte_cdk.test.declarative.test_suites.connector_base import MANIFEST_YAML
+from airbyte_cdk.test.standard_tests.connector_base import MANIFEST_YAML
 
 SAMPLE_COMPONENTS_PY_TEXT = """
 def sample_function() -> str:
